requirements.txt
av==12.0.0
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
tqdm==4.66.2
reedsolo==1.7.0